--- HuskStandalone/HuskStandalone.options
@@ -1,121 +1,0 @@
-[SceneFile]
-Type=filename
-Label=Scene Filename
-Category=Scene File
-Index=0
-Description=The scene filename as it exists on the network
-Required=true
-DisableIfBlank=true
-
-[OutputFile]
-Type=filenameSave
-Label=Output
-Category=Rendering Options
-Index=0
-Description=The output filename path
-Required=false
-DisableIfBlank=true
-
-[Threads]
-Type=integer
-Minimum=0
-Maximum=64
-Label=Threads
-Category=Rendering Options
-Index=1
-Description=The number of rendering threads (0 for system processor count)
-Required=false
-DisableIfBlank=true
-
-[Build]
-Type=enum
-Values=None;32bit;64bit
-Label=Build To Force
-Category=Rendering Options
-Index=2
-Description="for USD we are unsure whether we need this?"
-Required=false
-DisableIfBlank=true
-
-[Renderer]
-Type=label
-Label=Renderer
-Category=Rendering
-Index=0
-Description=The USD Hydra Delegate renderer to use for rendering.
-Required=false
-DisableIfBlank=true
-
-[RenderSettings]
-Type=label
-Label=Render Settings Prim Path
-Category=Rendering
-Index=0
-Description=The Render Settings Prim Path to use as Render Settings, defaults to "/Render/rendersettings" when not set.
-Required=false
-DisableIfBlank=true
-
-[Purpose]
-Type=label
-Label=Purpose
-Category=Rendering
-Index=0
-Description=The USD Hydra Delegate renderer to use for rendering.
-Required=false
-DisableIfBlank=true
-
-[Complexity]
-Type=label
-Label=Complexity
-Category=Rendering
-Index=0
-Description=The USD Hydra Delegate renderer to use for rendering.
-Required=false
-DisableIfBlank=true
-
-[Snapshot]
-Type=integer
-Minimum=-1
-Maximum=3600
-Label=Snapshot
-Category=Rendering
-Index=0
-Description=Save a snapshot of the current state of the rendered image every 'x' seconds. -1 means no snapshots. Default is -1.
-Required=false
-DisableIfBlank=true
-
-[PreRender]
-Type=filename
-Label=PreRender
-Category=Rendering
-Index=0
-Description=Pre-Render script to run.
-Required=false
-DisableIfBlank=true
-
-[PreFrame]
-Type=filename
-Label=PreFrame
-Category=Rendering
-Index=0
-Description=Pre-frame script to run.
-Required=false
-DisableIfBlank=true
-
-[PostFrame]
-Type=filename
-Label=PostFrame
-Category=Rendering
-Index=0
-Description=Post-frame script to run.
-Required=false
-DisableIfBlank=true
-
-[PostRender]
-Type=filename
-Label=PostRender
-Category=Rendering
-Index=0
-Description=Post-Render script to run.
-Required=false
-DisableIfBlank=true
